FlambergeAPI/Content/requirements.txt
Distance==0.1.3
fastapi==0.104.1
kmodes==0.12.2
numpy>=1.23.5
pandas==2.1.1
psycopg2-binary
scikit-learn==1.3.1